+++ resources/views/pages/partials/blocks/feature-grid.blade.php
@@ -1,0 +1,7 @@
+@php
+    $delegate = $block->replicate();
+    $delegate->setRelation('children', $block->children);
+    $delegate->variant = 'cards';
+@endphp
+
+@include('pages.partials.blocks.columns', ['block' => $delegate])