+++ resources/views/pages/partials/blocks/testimonial.blade.php
@@ -1,0 +1,7 @@
+@php
+    $delegate = $block->replicate();
+    $delegate->setRelation('children', $block->children);
+    $delegate->variant = 'testimonial';
+@endphp
+
+@include('pages.partials.blocks.quote', ['block' => $delegate])